% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO.simu.R
\name{CFO.simu}
\alias{CFO.simu}
\title{Find the maximum tolerated dose (MTD) for a single Calibration-Free Odds (CFO) trial.}
\usage{
CFO.simu(phi, p.true, ncohort, init.level=1, cohortsize=3, 
                alp.prior = phi, bet.prior = 1 - phi, seed=100)
}
\arguments{
\item{phi}{the target DLT rate}

\item{p.true}{the true DLT rates under the different dose levels}

\item{ncohort}{the total number of cohorts}

\item{init.level}{the dose level assigned to the first cohort. The default value \code{init.level} is 1.}

\item{cohortsize}{the sample size in each cohort}

\item{alp.prior, bet.prior}{the parameters of the prior distribution for the true DLT rate at any dose level.
This prior distribution is set to Beta(\code{alpha.prior}, \code{beta.prior}).
The default value is \code{phi} and \code{1-phi}.}

\item{seed}{the random seed for simulation}
}
\value{
The \code{CFO.next()} function returns a list object comprising the following components: the target DLT
rate ($target), the actual DLT rates under different dose levels ($p.true), the selected MTD ($MTD),
the total number of DLTs and patients for all dose levels ($DLT.ns and $dose.ns), and the over-toxicity status
for all dose levels ($over.doses). Specifically, the value of 1 represents over-toxicity at that dose level,
while the value of 0 indicates safety at that dose level.
}
\description{
Use this function to find the maximum tolerated dose (MTD) for a single Calibration-Free Odds (CFO) trial.
}
\details{
The \code{CFO.simu()} function is designed to determine the Maximum Tolerated Dose (MTD) for a single CFO trial.
Given the toxicity outcomes from previous cohorts, each cohort is sequentially assigned to the most suitable dose
level based on the decision rule. Early stopping criteria are incorporated into the CFO design to ensure patient
safety and benefit. If there is substantial evidence indicating that the current dose level exhibits excessive
toxicity (\eqn{\Pr(p_C > \phi|x_C, m_C \geq 3) > 0.95}), we exclude the current dose level as well as higher dose
levels from the trial. The trial will be terminated if the lowest dose level is overly toxicity. \cr
Upon reaching the predefined maximum sample size or satisfying the early stopping criteria, the experiment is
concluded, and the MTD is determined using isotonic regression.
}
\examples{
## find the MTD for a single CFO trial
phi <- 0.2; ncohort <- 12; cohortsize <- 3
p.true <-c(0.01, 0.05, 0.10, 0.14, 0.20, 0.26, 0.34)
CFO.simu(phi, p.true, ncohort, init.level=1, cohortsize=3, 
         alp.prior = phi, bet.prior = 1 - phi, seed = 100)
}
\references{
Jin, H., & Yin, G. (2022). CFO: Calibration-free odds design for phase I/II clinical trials.
\emph{Statistical Methods in Medical Research}, 31(6), 1051-1066.
}
\author{
Jialu Fang and Wenliang Wang
}
