% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO2d.oc.R
\name{CFO2d.oc}
\alias{CFO2d.oc}
\title{This function runs multiple simulations of the 2dCFO design and averages the operating characteristics.}
\usage{
CFO2d.oc(
  target,
  p.true,
  ncohort = 20,
  cohortsize = 3,
  init.level = c(1, 1),
  prior.para = list(alp.prior = target, bet.prior = 1 - target),
  seeds = NULL,
  nsimu = 1000
)
}
\arguments{
\item{target}{the target DLT rate.}

\item{p.true}{a matrix representing the true DIL rates under the different dose levels.}

\item{ncohort}{the total number of cohorts, the default value is 20.}

\item{cohortsize}{the number of patients or size of each cohort. the default value is 3.}

\item{init.level}{a numeric vector of length 2 representing the initial dose level (default is c(1,1)).}

\item{prior.para}{the prior parameters for a beta distribution, usually set as list(alp.prior=target, bet.prior=1-target) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}

\item{seeds}{A vector of random seed for each simulations, for example, \code{seeds = 1:nsimu} (default is NULL).}

\item{nsimu}{the total number of trials to be simulated. The default value is 1000.}
}
\value{
A list with the averaged operating characteristics across all simulations.
\itemize{
\item{p.true}{The matrix of the true DLT rates under the different dose levels.}
\item{selPercent}{The matrix of the selection percentage of each dose level}
\item{dose.ns}{A matrix of the averaged number of patients allocated for different doses in one simulation.}
\item{DLT.ns}{A matrix of the averaged number of DLT observed for different doses in one simulation.}
\item{pcorrect}{The percentage of the correct selection of the MTD.}
\item{npercent}{The averaged percentage of patients assigned to the target DLT rate.}
\item{ptoxic}{The averaged percentage of patients assigned to dose levels with a DLT rate greater than the target.}
\item{ntox}{The averaged total number of DLTs observed.}
}
}
\description{
This function runs multiple simulations of the 2dCFO design and averages the operating characteristics.
}
\examples{
## Simulate a two-dimensional dose-finding trial with 20 cohorts of size 3 for 1000 replications.
p.true <- matrix(c(0.05, 0.10, 0.15, 0.30, 0.45,
0.10, 0.15, 0.30, 0.45, 0.55,
0.15, 0.30, 0.45, 0.50, 0.60), 
nrow = 3, ncol = 5, byrow = TRUE)

CFO2d.oc(target=0.3, p.true=p.true, ncohort = 20, cohortsize = 3, seeds = 1:1000, nsimu = 1000)
}
