% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO2d.plot.R
\name{CFO2d.plot}
\alias{CFO2d.plot}
\title{Plotting Dose Function}
\usage{
CFO2d.plot(CFO2d.res)
}
\arguments{
\item{CFO2d.res}{A list result from the CFO2d.sim function. It should include the following components:
\itemize{
\item{dose}{A matrix of doses given to each cohort.}
\item{DLT}{A vector of DLTs observed in each cohort.}
\item{dose.ns}{A vector of the number of subjects at each dose.}
\item{p.true}{A matrix representing the true probabilities of DLT at each dose combination.}
}}
}
\value{
A ggplot object of the dose assigned to each patient, and whether DLT was observed.
The x-axis represents the sequence of patients treated, while the y-axis represents the combined dose level.
The points are filled if a DLT is observed.
}
\description{
This function takes the result of CFO2d.sim simulation and plots the combined dose level
for each sequence of patients treated. The doses are displayed as a step plot,
with points representing each patient. The point is filled if a dose-limiting toxicity (DLT) is observed.
}
\examples{
p.true <- matrix(c(0.05, 0.10, 0.15, 0.30, 0.45,
                   0.10, 0.15, 0.30, 0.45, 0.55,
                   0.15, 0.30, 0.45, 0.50, 0.60), 
                 nrow = 3, ncol = 5, byrow = TRUE)
CFO2d.res <- CFO2d.sim(phi=0.3, p.true, ncohort = 20, cohortsize = 3)
CFO2d.plot(CFO2d.res)
}
