% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO2d.selectmtd.R
\name{CFO2d.selectmtd}
\alias{CFO2d.selectmtd}
\title{Select the maximum tolerated dose (MTD) for real single drug trials}
\usage{
CFO2d.selectmtd(target, npts, ntox, prior.para=list(alp.prior=target, bet.prior=1-target), 
       cutoff.eli=0.95, extrasafe=FALSE, offset=0.05, verbose=TRUE)
}
\arguments{
\item{target}{the target DLT rate}

\item{npts}{a matrix containing the number of patients treated at each dose level}

\item{ntox}{a matrix containing the number of patients who experienced DLT at each dose level}

\item{prior.para}{the prior parameters for a beta distribution, usually set as list(alp.prior=target, bet.prior=1-target)
by default. \code{alp.prior} and \code{bet.prior} represent the parameters of the prior distribution for
the true DLT rate at any dose level. This prior distribution is specified as Beta( \code{alpha.prior},
\code{beta.prior}).}

\item{cutoff.eli}{the cutoff to eliminate overly toxic doses for safety. We recommend
the default value of (\code{cutoff.eli=0.95}) for general use.}

\item{extrasafe}{set \code{extrasafe=TRUE} to impose a more strict early stopping rule for
extra safety}

\item{offset}{a small positive number (between \code{0} and \code{0.5}) to control how strict the
stopping rule is when \code{extrasafe=TRUE}. A larger value leads to
a more strict stopping rule. The default value \code{offset=0.05}
generally works well.}

\item{verbose}{set \code{verbose=TRUE} to return more details of the results}
}
\value{
\code{CFO2d.selectmtd()} returns
\itemize{
\item{target: }{the target DLT rate.}
\item{MTD: }{the selected MTD.}
\item{p_est: }{the isotonic estimate of the DLT probablity at each dose and associated \eqn{95\%} credible interval.}
\item{p_est_CI: }{the credible interval for the isotonic estimate.}
}
}
\description{
Select the maximum tolerated dose (MTD) when the real single-drug trial is completed
}
\details{
\code{CFO2d.selectmtd()} selects the MTD based on isotonic estimates of toxicity
probabilities. \code{CFO2d.selectmtd()} selects as the MTD dose \eqn{j^*}, for which the
isotonic estimate of the DLT rate is closest to the target. If there
are ties, we select from the ties the highest dose level when the estimate
of the DLT rate is smaller than the target, or the lowest dose level
when the estimate of the DLT rate is greater than the target. The
isotonic estimates are obtained by the pooled-adjacent-violators algorithm
(PAVA) (Barlow, 1972).
}
\note{
The MTD selection and dose escalation/deescalation rule are two independent
components of the trial design. The MTD selection rule in the CFO package remains consistent with that in
the BOIN package; isotonic regression is employed to select the MTD after the completion of the trial.
When appropriate, another dose selection procedure (e.g., based on a fitted logistic model) can be used to select
the MTD after the completion of the trial using the CFO-type design.
}
\examples{

cns <- matrix(c(3, 3, 0,
                0, 6, 0,
                0, 0, 0), 
              nrow = 3, ncol = 3, byrow = TRUE)

cys <- matrix(c(0, 1, 0,
                0, 2, 0,
                0, 0, 0), 
              nrow = 3, ncol = 3, byrow = TRUE)
selmtd <- CFO2d.selectmtd(target=0.3, npts=cns, ntox=cys)
summary(selmtd)
plot(selmtd)
}
\references{
Jin, H., & Yin, G. (2022). CFO: Calibration-free odds design for phase I/II clinical trials.
\emph{Statistical Methods in Medical Research}, 31(6), 1051-1066. \cr
Liu S. and Yuan, Y. (2015). Bayesian Optimal Interval Designs for
Phase I Clinical Trials, \emph{Journal of the Royal Statistical Society:
Series C}, 64, 507-523. \cr
Yan, F., Zhang, L., Zhou, Y., Pan, H., Liu, S. and Yuan, Y. (2020).BOIN: An R Package
for Designing Single-Agent and Drug-Combination Dose-Finding Trials Using Bayesian Optimal
Interval Designs. \emph{Journal of Statistical Software}, 94(13), 1-32. \cr
Yuan Y., Hess K.R., Hilsenbeck S.G. and Gilbert M.R. (2016). Bayesian Optimal Interval Design: A
Simple and Well-performing Design for Phase I Oncology Trials, \emph{Clinical Cancer Research}, 22, 4291-4301. \cr
}
\author{
Jialu Fang, Wenliang Wang
}
