% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateonset.simu.R
\name{lateonset.simu}
\alias{lateonset.simu}
\title{Find the maximum tolerated dose (MTD) for a single trial using the CFO-type designs with late-onset toxicities.}
\usage{
lateonset.simu(target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
       accrual.dist, design, init.level = 1, 
       prior.para = list(alp.prior = target, bet.prior = 1 - target), 
       seed = NULL, cutoff.eli = 0.95, extrasafe = FALSE, offset = 0.05)
}
\arguments{
\item{target}{the target DLT rate.}

\item{p.true}{the true DLT rates under the different dose levels.}

\item{tau}{maximal assessment window size.}

\item{cohortsize}{the number of patients or size of each cohort. The default value \code{cohortsize} is 3.}

\item{ncohort}{the total number of cohorts.}

\item{tte.para}{the parameter related with the distribution of the time to DLT events. The time to DLT is sample from a Weibull
distribution, with \code{tte.para} representing the proportion of DLTs occurring within the initial half of the
assessment window \code{tau}.}

\item{accrual}{the accrual rate, i.e., the number of patients accrued in per unit time.}

\item{accrual.dist}{the distribution of the arrival times of patients. When \code{accrual.dist = 'fix'}, it corresponds to all
patients in each cohort arriving simultaneously at a given accrual rate. When \code{accrual.dist = 'unif'},
it corresponds to a uniform distribution, and when \code{accrual.dist = 'exp'}, it corresponds to an
exponential distribution.}

\item{design}{option for selecting different designs, which can be set as \code{'TITE-CFO'}, \code{'TITE-aCFO'},
\code{'fCFO'}, \code{'f-aCFO'}, \code{'bCFO'}, and \code{'b-aCFO'}. Specifically, \code{'bCFO'} refers
to the benchmark CFO, and \code{'b-aCFO'} denotes the benchmark aCFO.}

\item{init.level}{the dose level assigned to the first cohort. The default value \code{init.level} is 1.}

\item{prior.para}{the prior parameters for a beta distribution, usually set as list(alp.prior = target, bet.prior = 1 - target) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}

\item{seed}{an integer to set as the seed of the random number generator for reproducible results, the default is set to NULL.}

\item{cutoff.eli}{the cutoff to eliminate overly toxic doses for safety. We recommend
the default value of (\code{cutoff.eli = 0.95}) for general use.}

\item{extrasafe}{set \code{extrasafe = TRUE} to impose a more strict early stopping rule for
extra safety.}

\item{offset}{a small positive number (between \code{0} and \code{0.5}) to control how strict the
stopping rule is when \code{extrasafe=TRUE}. A larger value leads to
a more strict stopping rule. The default value \code{offset = 0.05}
generally works well.}
}
\value{
The \code{lateonset.simu()} function returns a list object comprising the following components:
\itemize{
\item{target: }{the target DLT rate.}
\item{MTD: }{the selected MTD. \code{MTD=99} indicates that this trial is terminated due to early stopping.}
\item{correct: }{a binary indicator of whether the recommended dose level matches the target DLT rate (1 for yes).}
\item{npatients: }{the total number of patients allocated for all dose levels}
\item{ntox: }{the total number of DLTs observed for all dose levels.}
\item{npercent: }{the percentage of subjects assigned to the target DLT rate.}
\item{over.doses: }{a vector indicating whether each dose is overdosed or not (1 for yes).}
\item{cohortdose: }{a vector including the dose level assigned for each cohort.}
\item{patientDLT: }{a vector including the DLT outcome observed for each patient.}
\item{totaltime: }{the duration of the trial.}
\item{entertimes: }{the time that each participant enters the trial.}
\item{DLT.times: }{the time to DLT for each subject in the trial. If no DLT occurs for a certain subject,
\code{DLT.times} is 0.}
}
}
\description{
Use this function to find the maximum tolerated dose (MTD) for the CFO-type designs with late-onset toxicities,
specifically, including Time-to-event CFO (TITE-CFO) design, fractional CFO (fCFO), benchmark CFO, TITE-aCFO design,
the f-aCFO design and benchmark aCFO design.
}
\note{
The results returned by \code{lateonset.simu()} represent a single trial outcome. In practical
applications, multiple simulations are required to ascertain the MTD, and this can be achieved using the
\code{CFO.oc()} function.
}
\examples{
target <- 0.2; ncohort <- 12; cohortsize <- 3
p.true <- c(0.02, 0.05, 0.20, 0.28, 0.34, 0.40, 0.44)
tau <- 3; accrual <- 2; tte.para <- 0.5; accrual.dist <- 'unif'
## find the MTD for a single TITE-CFO trial
TITECFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                accrual.dist, design = 'TITE-CFO', init.level = 1, 
                prior.para = list(alp.prior = target, bet.prior = 1 - target), seed = 1)
summary(TITECFOtrial)
plot(TITECFOtrial)
## find the MTD for a single TITE-aCFO trial
TITEaCFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                accrual.dist, design = 'TITE-aCFO', init.level = 1, 
                prior.para = list(alp.prior = target, bet.prior = 1 - target), seed = 1)
summary(TITEaCFOtrial)
plot(TITEaCFOtrial)
## find the MTD for a single fCFO trial
fCFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                accrual.dist, design = 'fCFO', init.level = 1, 
                prior.para = list(alp.prior = target, bet.prior = 1 - target), seed = 1)
summary(fCFOtrial)
plot(fCFOtrial)
## find the MTD for a single f-aCFO trial
faCFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                accrual.dist, design = 'f-aCFO', init.level = 1, 
                prior.para = list(alp.prior = target, bet.prior = 1 - target), seed = 1)
summary(faCFOtrial)
plot(faCFOtrial)
## find the MTD for a single benchmark CFO trial
bCFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                accrual.dist, design = 'bCFO', init.level = 1, 
                prior.para = list(alp.prior = target, bet.prior = 1 - target), seed = 1)
summary(bCFOtrial)
plot(bCFOtrial)
## find the MTD for a single benchmark aCFO trial
baCFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                accrual.dist, design = 'b-aCFO', init.level = 1, 
                prior.para = list(alp.prior = target, bet.prior = 1 - target), seed = 1)
summary(baCFOtrial)
plot(baCFOtrial)
}
\references{
Jin, H., & Yin, G. (2022). CFO: Calibration-free odds design for phase I/II clinical trials.
\emph{Statistical Methods in Medical Research}, 31(6), 1051-1066. \cr
Jin, H., & Yin, G. (2023). Time‐to‐event calibration‐free odds design: A robust efficient design for
phase I trials with late‐onset outcomes. \emph{Pharmaceutical Statistics}. \cr
Yin, G., Zheng, S., & Xu, J. (2013). Fractional dose-finding methods with late-onset toxicity in
phase I clinical trials. \emph{Journal of Biopharmaceutical Statistics}, 23(4), 856-870.
}
\author{
Jialu Fang
}
