% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCFO.next.R
\name{aCFO.next}
\alias{aCFO.next}
\title{Determination of the dose level for next cohort in aCFO design}
\usage{
aCFO.next(phi, tys, tns, alp.prior, bet.prior, tover.doses, cidx)
}
\arguments{
\item{phi}{the target DLT rate}

\item{tys}{the current number of DLTs observed in patients for all dose levels.}

\item{tns}{the current number of patients for all dose levels.}

\item{alp.prior, bet.prior}{the parameters of the prior distribution for the true DLT rate at any dose level.
This prior distribution is set to Beta( \code{alpha.prior}, \code{beta.prior}).
The default value is \code{phi} and \code{1-phi}.}

\item{tover.doses}{whether the dose level (from the first to last dose level) is over-toxic or not.
The value is set as 1 if the dose level is overly toxicity; otherwise, it is set to 0.}

\item{cidx}{dose level for current cohort}
}
\value{
The \code{aCFO.next()} function returns a list object comprising the following elements: the target DLT
rate ($target), the current number of DLTs and patients for the left, current, and right dose levels ($cys and $cns),
the decision in the aCFO design of whether to move to the left or right dose level for the next cohort ($decision), and the
corresponding index ($index). Specifically, 1 corresponds to de-escalation, 2 corresponds to staying at the
current dose, and 3 corresponds to escalation.
}
\description{
In aCFO design, use the function to determine the dose movement based on the toxicity outcomes of the enrolled cohorts.
}
\details{
The aCFO design design incorporate the dose information of all positions (from the lowest to the
highest dose levels) into the trial decision-making. Prior to assigning dose levels for new patient
cohorts, aCFO compares the evidence from the current dose level with all doses to its left and right.
This design is rooted in the odds ratio, specifically \eqn{O_C / \overline{O}_{J}} and
\eqn{\overline{O}_C / O_R} from the CFO design. By aggregating odds ratios from the left and right sides,
it forms two collective statistics: \eqn{ {\rm OR}_L =\sum_{i=1}^{J} O_C/ \overline{O}_{L_i} } for dose
de-escalation (movement to the left) and \eqn{ {\rm OR}_R = \sum_{i=1}^{H} \overline{O}_C / O_{R_i} }
for dose escalation (movement to the right), where J and H represent the counts of doses on the left and
right sides of the current dose, respectively. For the new statistic \eqn{ {\rm OR}_L } and \eqn{ {\rm OR}_R },
their corresponding thresholds are derived by summing up its individual thresholds \eqn{\gamma_{L_i}} and
\eqn{\gamma_{R_i}}, i.e., \eqn{\sum_{i=1}^{J}\gamma_{L_i}} and \eqn{\sum_{i=1}^{H}\gamma_{R_i}}. \cr
Besides，The aCFO design retains the same early stopping criteria as the CFO design. Overall，while preserving
the nature of the CFO design (model-free and calibration-free), the aCFO designs enhance the efﬁciency by
incorporating more dose information.
}
\examples{
## determine the dose level for the next cohort of new patients
tys <- c(0,0,1,0,0,0,0); tns <- c(3,3,6,0,0,0,0)
aCFO.next(phi=0.2, tys=tys, tns=tns, alp.prior=0.2, 
          bet.prior=0.8, tover.doses=c(0,0,0,0,0,0,0), cidx=3)

}
\references{
Jin, H., & Yin, G. (2022). CFO: Calibration-free odds design for phase I/II clinical trials.
\emph{Statistical Methods in Medical Research}, 31(6), 1051-1066.
}
\author{
Jialu Fang
}
