% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO.next.R
\name{CFO.next}
\alias{CFO.next}
\title{Determination of the dose level for next cohort}
\usage{
CFO.next(phi, cys, cns, curDose, add.args=list(alp.prior=phi,bet.prior=1-phi))
}
\arguments{
\item{phi}{the target DLT rate}

\item{cys}{the current number of DLTs observed at the left, current, and right dose levels.}

\item{cns}{the current number of patients treated at the left, current, and right dose levels.}

\item{curDose}{the current dose level.}

\item{add.args}{additional parameters, usually set as list(alp.prior=phi, bet.prior=1-phi) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}
}
\value{
The \code{CFO.next()} function returns a list object comprising the following elements:
\itemize{
\item{taget: }{the target DLT rate.}
\item{cys: }{the current counts of DLTs observed at the left, current, and right dose levels}
\item{cns: }{the current counts of patients treated at the left, current, and right dose levels}
\item{decision: }{the decision in the CFO design, where \code{left}, \code{stay}, and \code{right} represent the
movement directions, and \code{stop} indicates stopping the experiment}
\item{curDoses: }{the current level.}
\item{nextDose: }{the recommended dose level for the next cohort.}
\item{overTox: }{the situation regarding which position experiences overly toxicity, where 'NA' signifies that the
occurrence of overly toxicity did not happen.}
}
}
\description{
Use the function to determine the dose movement based on the toxicity outcomes of the enrolled cohorts.
}
\details{
The CFO design determines the dose level for the next cohort by assessing evidence from the current
dose level and its adjacent levels. This evaluation is based on odds ratios denoted as \eqn{O_k}, where
k = L, C, R represents left, current, and right dose levels. Additionally, we define \eqn{\overline{O}_k = 1/O_k}.
The ratio \eqn{O_C / \overline{O}_{L}} indicates the inclination for de-escalation, while \eqn{\overline{O}_C / O_R}
quantifies the tendency for escalation. Threshold values \eqn{\gamma_L} and \eqn{\gamma_R} are chosen to
minimize the probability of making incorrect decisions.The decision process is summarized in Table 1
of Jin and Yin (2022).
An overdose control rule is implemented to ensure patient safety. If the data suggest excessive
toxicity at the current dose level, we exclude that level and those higher levels. Two scenarios
lead to a decision on one side only: when the current dose is at the boundary (the first or last dose level)
or when higher dose levels have been eliminated.
}
\note{
When the current dose level is the lowest or highest (i.e., at the boundary), the parts in \code{cys} and
\code{cns} where there is no data are filled with NA.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     The position indicated by \code{overTox} experience overly toxicity. In the complete single trial, the dose 
     level and all the dose levels above will be eliminated.
}\if{html}{\out{</div>}}
}
\examples{
## determine the dose level for the next cohort of new patients
cys <- c(0,1,0); cns <- c(3,6,0)
CFO.next(phi=0.2, cys=cys, cns=cns, curDose=3, add.args=list(alp.prior=0.2, bet.prior=0.8))

cys <- c(NA,3,0); cns <- c(NA,3,0)
CFO.next(phi=0.2, cys=cys, cns=cns, curDose=1, add.args=list(alp.prior=0.2, bet.prior=0.8))

cys <- c(0,3,NA); cns <- c(3,3,NA)
CFO.next(phi=0.2, cys=cys, cns=cns, curDose=7, add.args=list(alp.prior=0.2, bet.prior=0.8))

}
\references{
Jin, H., & Yin, G. (2022). CFO: Calibration-free odds design for phase I/II clinical trials.
\emph{Statistical Methods in Medical Research}, 31(6), 1051-1066.
}
\author{
Jialu Fang and Wenliang Wang
}
