% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO.oc.R
\name{CFO.oc}
\alias{CFO.oc}
\title{Generate operating characteristics for multiple simulations}
\usage{
CFO.oc(nsimu=5000, design, target, p.true, ncohort, init.level, cohortsize,
       tau=NA, accrual=NA, tite.dist=NA, accrual.dist=NA, 
       prior.para=list(alp.prior=target, bet.prior=1-target), seeds = NULL,
       cutoff.eli=0.95, extrasafe=FALSE, offset=0.05)
}
\arguments{
\item{nsimu}{the total number of trials to be simulated. The default value is 5000.}

\item{design}{option for selecting different designs, which can be set as \code{'CFO'}, \code{'aCFO'},
\code{'TITE-CFO'}, \code{'TITE-aCFO'}, \code{'fCFO'}, \code{'f-aCFO'}, \code{'bCFO'},
and \code{'b-aCFO'}. Specifically, \code{'bCFO'} refers to the benchmark CFO, and
\code{'b-aCFO'} denotes the benchmark aCFO.}

\item{target}{the target DLT rate.}

\item{p.true}{the true DLT rates under the different dose levels.}

\item{ncohort}{the total number of cohorts.}

\item{init.level}{the dose level assigned to the first cohort. The default value \code{init.level} is 1.}

\item{cohortsize}{the number of patients or size of each cohort. The default value \code{cohortsize} is 3.}

\item{tau}{maximal assessment window size. 'NA' should be assigned if the design without late-oneset outcomes.}

\item{accrual}{the accrual rate, i.e., the number of patients accrued in tau time. 'NA' should be assigned
if the design without late-oneset outcomes.}

\item{tite.dist}{the distribution of the time to DLT events. \code{tite.dist=1} corresponds to a uniform distribution,
\code{tite.dist=2} corresponds to a Weibull distribution, and \code{tite.dist=3} corresponds to a
log-logistic distribution. 'NA' should be assigned if the design without late-oneset outcomes.}

\item{accrual.dist}{the distribution of the arrival times of patients. When \code{tite.dist=1}, it corresponds to all
patients in each cohort arriving simultaneously at a given accrual rate. When \code{tite.dist=2},
it corresponds to a uniform distribution, and when \code{tite.dist=3}, it corresponds to an
exponential distribution. 'NA' should be assigned if the design without late-oneset outcomes.}

\item{prior.para}{the prior parameters for a beta distribution, usually set as list(alp.prior=target, bet.prior=1-target) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}

\item{seeds}{A vector of random seeds for each simulation, for example, \code{seeds = 1:nsimu} (default is NULL).}

\item{cutoff.eli}{the cutoff to eliminate overly toxic doses for safety. We recommend
the default value of (\code{cutoff.eli=0.95}) for general use.}

\item{extrasafe}{set \code{extrasafe=TRUE} to impose a more strict early stopping rule for
extra safety}

\item{offset}{a small positive number (between \code{0} and \code{0.5}) to control how strict the
stopping rule is when \code{extrasafe=TRUE}. A larger value leads to
a more strict stopping rule. The default value \code{offset=0.05}
generally works well.}
}
\value{
the \code{ CFO.oc() } function returns basic setup of ($simu.setup and $p.true) and the operating
characteristics ($selPercent,$dose.ns,$DLT.ns and $simu.oc) of the design. \cr
The operating characteristics includes \cr
(1) selection percentage at each dose level ($selPercent) \cr
(2) averaged number of toxicity observed at each dose level in one simulation ($DLT.ns) \cr
(3) averaged number of patients treated at each dose level in one simulation ($dose.ns) \cr
(4) percentage of correct selection of the MTD ($simu.oc$MTDSel) \cr
(5) percentage of patients allocated to the MTD ($simu.oc$MTDAllo) \cr
(6) percentage of selecting a dose above the MTD ($simu.oc$overSel) \cr
(7) percentage of allocating patients at dose levels above the MTD ($simu.oc$overAllo) \cr
(8) percentage of the patients suffering DLT ($simu.oc$averDLT) \cr
(9) average trial duration if trials with late-onset toxicities ($simu.oc$averDur) \cr
(10) number of simulations terminated due to early stopping ($simu.oc$errStop) \cr
(11) number of patients entering the trial ($simu.oc$tol.Subjs)
}
\description{
Obtain the operating characteristics of the CFO-type designs for multiple simulations.
}
\details{
The operating characteristics of CFO-type and aCFO-type designs are generated by simulating trials under the
pre-specified true toxicity probabilities of the investigational doses. The choice of which design to execute
is determined by setting the \code{design} parameter. Some parameters (\code{tau}, \code{accrual},
\code{tite.dist}, and \code{accrual.dist}) need to be set only when running a design that can handle late-onset
toxicities; otherwise, they default to NA.\cr
Additionally, in the example, we set \code{nsimu=100} for testing time considerations. In reality, \code{nsimu}
is typically set to 5000 to ensure the accuracy of the results.
}
\examples{
## setting
nsimu <- 100; target <- 0.2; ncohort <- 12; cohortsize <- 3; init.level <- 1
p.true <-c(0.01, 0.05, 0.10, 0.14, 0.20, 0.26, 0.34)
prior.para=list(alp.prior=target, bet.prior=1-target)
tau <- 3; accrual <- 6; tite.dist <- 2; accrual.dist <- 2

## get the operating characteristics for 100 simulations using the CFO design
CFOoc <- CFO.oc (nsimu, design='CFO', target, p.true, ncohort, init.level, cohortsize,
         tau=NA, accrual=NA, tite.dist=NA, accrual.dist=NA, prior.para, seeds = 1:nsimu)
summary(CFOoc)
plot(CFOoc)
## get the operating characteristics for 100 simulations using the aCFO design
aCFOoc <- CFO.oc (nsimu, design='aCFO', target, p.true, ncohort, init.level, cohortsize,
       tau=NA, accrual=NA, tite.dist=NA, accrual.dist=NA, prior.para, seeds = 1:nsimu)
summary(aCFOoc)
plot(aCFOoc)
## get the operating characteristics for 100 simulations using the TITE-CFO design
TITECFOoc <- CFO.oc (nsimu, design='TITE-CFO', target, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, prior.para, seeds = 1:nsimu)
summary(TITECFOoc)
plot(CFOoc)
## get the operating characteristics for 100 simulations using the TITE-aCFO design
TITEaCFOoc <- CFO.oc (nsimu, design='TITE-aCFO', target, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, prior.para)
summary(TITEaCFOoc)
plot(TITEaCFOoc)
## get the operating characteristics for 100 simulations using the fCFO design
fCFOoc <- CFO.oc (nsimu, design='fCFO', target, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, prior.para, seeds = 1:nsimu)
summary(fCFOoc)
plot(fCFOoc)
## get the operating characteristics for 100 simulations using the f-aCFO design
faCFOoc <- CFO.oc (nsimu, design='f-aCFO', target, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, prior.para, seeds = 1:nsimu)
summary(faCFOoc)
plot(faCFOoc)
## get the operating characteristics for 100 simulations using the bCFO design
bCFOoc <- CFO.oc (nsimu, design='bCFO', target, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, prior.para, seeds = 1:nsimu)
summary(bCFOoc)
plot(bCFOoc)
## get the operating characteristics for 100 simulations using the b-aCFO design
baCFOoc <- CFO.oc (nsimu, design='b-aCFO', target, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, prior.para, seeds = 1:nsimu)
summary(baCFOoc)
plot(baCFOoc)
}
