% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO2d.next.R
\name{CFO2d.next}
\alias{CFO2d.next}
\title{CFO2d.next function}
\usage{
CFO2d.next(
  target,
  cys,
  cns,
  currdose,
  prior.para = list(alp.prior = target, bet.prior = 1 - target),
  seed = NULL
)
}
\arguments{
\item{target}{The target DLT rate.}

\item{cys}{A matrix of the number of DLTs observed for each dose combination.}

\item{cns}{A matrix of the number of patients allocated for each dose combination.}

\item{currdose}{A vector of the current dose indices in the horizontal and vertical direction.}

\item{prior.para}{the prior parameters for a beta distribution, usually set as list(alp.prior=target, bet.prior=1-target) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}

\item{seed}{an integer to be set as the seed of the random number generator for reproducible results.}
}
\value{
A list with the following components:
\itemize{
\item{target}{The target DLT rate.}
\item{cys}{A 3 by 3 matrix of the number of DLT observed for each dose combination at and around the current dose.}
\item{cns}{A 3 by 3 matrix of the number of patients allocated for each dose combination at and around the current dose.}
\item{index}{A vector of length 2 representing the decision index for vertical and horizontal directions, where -1, 0, 1 represent de-escalation, stay and escalation respectively. }
\item{decision}{A vector of length 2 representing the recommended decisions for vertical and horizontal directions. }
\item{target}{The target DLT rate.}
\item{currdose}{The current dose combination.}
\item{nextdose}{The recommended dose combination for the next cohort.}
\item{overTox}{The dose levels that are considered overly toxic.}
}
}
\description{
This function is used to determine the next dose level in a 2dCFO design.
}
\examples{
cns <- matrix(c(3, 3, 0,
                0, 6, 0,
                0, 0, 0), 
              nrow = 3, ncol = 3, byrow = TRUE)

cys <- matrix(c(0, 1, 0,
                0, 2, 0,
                0, 0, 0), 
              nrow = 3, ncol = 3, byrow = TRUE)
currdose <- c(2,3)
CFO2d.next(target = 0.3, cys, cns, currdose = currdose, seed = 1)

}
\author{
Wenliang Wang
}
