% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO2d.next.R
\name{CFO2d.next}
\alias{CFO2d.next}
\title{CFO2d.next function}
\usage{
CFO2d.next(
  phi,
  cys,
  cns,
  cover.doses = matrix(0, 3, 3),
  curDose,
  add.args = list(alp.prior = phi, bet.prior = 1 - phi),
  seed = NULL
)
}
\arguments{
\item{phi}{The efficacy threshold.}

\item{cys}{A matrix of the number of successes for each dose combination.}

\item{cns}{A matrix of the number of patients for each dose combination.}

\item{cover.doses}{whether the dose level (the current dose and its adjacent doses) is over-toxic or not.
The value is set as 1 if the dose level is overly toxicity; otherwise, it is set to 0.}

\item{curDose}{A vector of the current dose indices in the horizontal and vertical direction.}

\item{add.args}{additional parameters, usually set as list(alp.prior=phi, bet.prior=1-phi) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}

\item{seed}{an integer to set as the seed of the random number generator for reproducible results.}
}
\value{
A list with the following components:
\itemize{
\item{target}{The target DLT rate.}
\item{cys}{A matrix of the number of DLT observed at the current dose and its adjacent doses.}
\item{cns}{A matrix of the number of patients treated at the current dose and its adjacent doses.}
\item{index}{A vector of length 2 representing the decision index for vertical and horizontal directions, where -1, 0, 1 represent de-escalation, stay and escalation respectively. }
\item{decision}{A vector of length 2 representing the recommended decisions for vertical and horizontal directions. }
\item{target}{The target DLT rate.}
\item{curDoses}{The current dose combination.}
\item{nextDose}{The recommended dose combination for the next cohort.}
}
}
\description{
This function is used to determine the next dose level in a 2dCFO design.
}
\examples{
cns <- matrix(c(3, 3, 0,
                0, 6, 0,
                0, 0, 0), 
              nrow = 3, ncol = 3, byrow = TRUE)

cys <- matrix(c(0, 1, 0,
                0, 2, 0,
                0, 0, 0), 
              nrow = 3, ncol = 3, byrow = TRUE)
curDose <- c(2,3)
CFO2d.next(0.3, cys, cns, curDose = curDose)

}
\author{
Wenliang Wang
}
