% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO2d.oc.R
\name{CFO2d.oc}
\alias{CFO2d.oc}
\title{This function runs multiple simulations of the 2dCFO design and averages the operating characteristics.}
\usage{
CFO2d.oc(
  phi,
  p.true,
  ncohort = 20,
  cohortsize = 3,
  init.level = c(1, 1),
  add.args = list(alp.prior = phi, bet.prior = 1 - phi),
  seed = NULL,
  n_sim = 1000
)
}
\arguments{
\item{phi}{Target dose-intensity level (DIL) rate.}

\item{p.true}{A matrix representing the true DIL rates under the different dose levels.}

\item{ncohort}{Integer. The number of cohorts (default is 20).}

\item{cohortsize}{Integer. The sample size in each cohort (default is 3).}

\item{init.level}{A numeric vector of length 2 representing the initial dose level (default is c(1,1)).}

\item{add.args}{An optional list of additional arguments, includes 'alp.prior' (default is phi) and 'bet.prior'(default is 1-phi).}

\item{seed}{A vector of random seed for each simulations, for example, \code{seeds = 1:n_sim} (default is NULL).}

\item{n_sim}{Integer. The number of simulations to run (default is 1000).}
}
\value{
A list with the averaged operating characteristics across all simulations.
\itemize{
\item{p.true}{The matrix of the true DLT rates under the different dose levels.}
\item{selPercent}{The matrix of the selection percentage of each dose level}
\item{dose.ns}{A matrix of the averaged number of patients allocated for different doses in one simulation.}
\item{DLT.ns}{A matrix of the averaged number of DLT observed for different doses in one simulation.}
\item{pcorrect}{The percentage of the correct selection of the MTD.}
\item{npercent}{The averaged percentage of patients assigned to the target DLT rate.}
\item{ptoxic}{The averaged percentage of patients assigned to dose levels with a DLT rate greater than the target.}
\item{ntox}{The averaged total number of DLTs observed.}
}
}
\description{
This function runs multiple simulations of the 2dCFO design and averages the operating characteristics.
}
\examples{
## Simulate a two-dimensional dose-finding trial with 20 cohorts of size 3 for 1000 replications.
p.true <- matrix(c(0.05, 0.10, 0.15, 0.30, 0.45,
0.10, 0.15, 0.30, 0.45, 0.55,
0.15, 0.30, 0.45, 0.50, 0.60), 
nrow = 3, ncol = 5, byrow = TRUE)

CFO2d.oc(phi=0.3, p.true=p.true, ncohort = 20, cohortsize = 3, n_sim = 100)
}
