% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO2d.simu.R
\name{CFO2d.simu}
\alias{CFO2d.simu}
\title{Obtain the operating characteristics of the 2dCFO designs for multiple simulations.}
\usage{
CFO2d.simu(target, p.true, ncohort=20, cohortsize=3, init.level=c(1,1), 
                 prior.para=list(alp.prior=target, bet.prior=1-target), seed=NULL)
}
\arguments{
\item{target}{the target DLT rate.}

\item{p.true}{a matrix representing the true DIL rates under the different dose levels.}

\item{ncohort}{the total number of cohorts, the default value is 20.}

\item{cohortsize}{the number of patients or size of each cohort. the default value is 3.}

\item{init.level}{the dose level assigned to the first cohort. The default value \code{init.level} is c(1,1).}

\item{prior.para}{the prior parameters for a beta distribution, usually set as list(alp.prior=target, bet.prior=1-target) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}

\item{seed}{an integer to be set as the seed of the random number generator for reproducible results, the default is set to NULL.}
}
\value{
A list with the following components:
\itemize{
\item{target: }{the target DLT rate.}
\item{MTD: }{a vector of length 2 representing the recommended dose level. \code{MTD=99} indicates that this trial is terminated due to early stopping.}
\item{correct: }{a binary indicator of whether the recommended dose level matches the target DLT rate (1 for yes).}
\item{npatients: }{a matrix of the number of patients allocated for different doses.}
\item{ntox: }{a matrix of the number of DLT observed for different doses.}
\item{npercent: }{the percentage of subjects assigned to the target DLT rate.}
\item{over.doses: }{a matrix indicating whether each dose is overdosed or not (1 for yes).}
\item{cohortdose: }{the dose combination assigned for each cohort.}
\item{patientDLT: }{the DLT observed at each cohort.}
}
}
\description{
Obtain the operating characteristics of the 2dCFO designs for multiple simulations.
}
\details{
The \code{CFO2d.simu} function simulates the operating characteristics of the CFO designs
in a two-dimensional dose-finding trial. The function uses parameters such as target DLT rate, true DLT rates
under different dose levels, and cohort details, and provides detailed output for performance evaluation.
It relies on the BOIN package.
}
\examples{
## Simulate a two-dimensional dose-finding trial with 20 cohorts of size 3.
p.true <- matrix(c(0.05, 0.10, 0.15, 0.30, 0.45,
                   0.10, 0.15, 0.30, 0.45, 0.55,
                   0.15, 0.30, 0.45, 0.50, 0.60), 
                 nrow = 3, ncol = 5, byrow = TRUE)

CFO2dtrial <- CFO2d.simu(target=0.3, p.true=p.true, ncohort = 20, cohortsize = 3, seed = 1)
summary(CFO2dtrial)
plot(CFO2dtrial)
}
\author{
Wenliang Wang
}
