% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cfo.R
\name{plot.cfo}
\alias{plot.cfo}
\title{Plot the simulation results for CFO-type and aCFO-type designs}
\usage{
\method{plot}{cfo}(x, ..., name = deparse(substitute(x)))
}
\arguments{
\item{x}{the object returned by other functions}

\item{...}{ignored arguments}

\item{name}{the name of the object to be plotted.
User doesn't need to input this parameter.}
}
\value{
\code{plot()} returns a figure or a series of figures depending on the object entered
}
\description{
Plot the objects returned by other functions, including (1) dose allocation of a single trial;
(2) operating characteristics of multiple simulations, including selesction percentage and the
number of patients treated at each dose
}
\details{
\code{plot()} returns a figure or a series of figures depending on the object entered.
Additionally, in the example, we set \code{nsimu=100} for testing time considerations.
In reality, \code{nsimu} is typically set to 5000 to ensure the accuracy of the results.
}
\examples{
############design without late-onset outcomes################
nsimu <- 100; phi <- 0.2; ncohort <- 12; cohortsize <- 3; init.level <- 1
p.true <-c(0.01, 0.05, 0.10, 0.14, 0.20, 0.26, 0.34)
add.args=list(alp.prior=phi, bet.prior=1-phi)
## CFO design
CFOtrial <- CFO.simu(phi, p.true, ncohort, init.level, cohortsize=3,add.args, accumulation = FALSE)
plot.cfo(CFOtrial)
CFOsimu <- CFO.oc (nsimu, design='CFO', phi, p.true, ncohort, init.level, cohortsize,
                    tau=NaN, accrual=NaN, tite.dist=NaN, accrual.dist=NaN, add.args)
plot.cfo(CFOsimu)
## aCFO design
aCFOtrial <- CFO.simu(phi, p.true, ncohort, init.level, cohortsize=3,add.args,
                    accumulation = TRUE)
plot.cfo(aCFOtrial)
aCFOsimu <- CFO.oc (nsimu, design='aCFO', phi, p.true, ncohort, init.level, cohortsize,
                    tau=NaN, accrual=NaN, tite.dist=NaN, accrual.dist=NaN, add.args)
plot.cfo(aCFOsimu)


##############design with late-onset outcomes################
tau <- 3; accrual <- 6; tite.dist <- 2; accrual.dist <- 1
## TITE-CFO design
TITECFOtrial <- lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist,
                accrual.dist, design='TITE-CFO', init.level, add.args)
plot.cfo(TITECFOtrial)
TITECFOsimu <- CFO.oc (nsimu, design='TITE-CFO', phi, p.true, ncohort, init.level, cohortsize,
                      tau, accrual, tite.dist, accrual.dist, add.args)
plot.cfo(TITECFOsimu)
## TITE-aCFO design
TITEaCFOtrial <- lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist,
                 accrual.dist,design='TITE-aCFO', init.level, add.args)
plot.cfo(TITEaCFOtrial)
TITEaCFOsimu <- CFO.oc (nsimu, design='TITE-aCFO', phi, p.true, ncohort, init.level, cohortsize,
                      tau, accrual, tite.dist, accrual.dist, add.args)
plot.cfo(TITEaCFOsimu)
## fCFO design
fCFOtrial <- lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist,
                accrual.dist, design='fCFO', init.level, add.args)
plot.cfo(fCFOtrial)
fCFOsimu <- CFO.oc (nsimu, design='fCFO', phi, p.true, ncohort, init.level, cohortsize,
                      tau, accrual, tite.dist, accrual.dist, add.args)
plot.cfo(fCFOsimu)
## f-aCFO design
faCFOtrial <- lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist,
                accrual.dist, design='f-aCFO', init.level, add.args)
plot.cfo(faCFOtrial)
faCFOsimu <- CFO.oc (nsimu, design='f-aCFO', phi, p.true, ncohort, init.level, cohortsize,
                      tau, accrual, tite.dist, accrual.dist, add.args)
plot.cfo(faCFOsimu)

##############two-dim design with late-onset outcomes################
p.true <- matrix(c(0.05, 0.10, 0.15, 0.30, 0.45,
                   0.10, 0.15, 0.30, 0.45, 0.55,
                   0.15, 0.30, 0.45, 0.50, 0.60), 
                 nrow = 3, ncol = 5, byrow = TRUE)

## plot the single trail returned by CFO2d.sim()
CFO2dtrail <- CFO2d.sim(phi=0.3, p.true=p.true, ncohort = 20, cohortsize = 3)
plot.cfo(CFO2dtrail)

## plot the multiple trails returned by CFO2d.oc()
CFO2dsim <- CFO2d.oc(phi=0.3, p.true=p.true, ncohort = 20, cohortsize = 3, n_sim = 100)
plot.cfo(CFO2dsim)

}
\author{
Jialu Fang
}
