% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cfo.R
\name{plot.cfo}
\alias{plot.cfo}
\title{Plot the simulation results for CFO-type and aCFO-type designs}
\usage{
\method{plot}{cfo}(x, ..., name = deparse(substitute(x)))
}
\arguments{
\item{x}{the object returned by other functions}

\item{...}{ignored arguments}

\item{name}{the name of the object to be plotted.
User doesn't need to input this parameter.}
}
\value{
\code{plot()} returns a figure or a series of figures depending on the object entered
}
\description{
Plot the objects returned by other functions, including (1) dose allocation of a single trial;
(2) operating characteristics of multiple simulations, including MTD selesction percentage,
the averaged number of patients allocated for different doses in one simulation and the averaged
number of DLT observed for different doses in one simulation.
}
\details{
\code{plot()} returns a figure or a series of figures depending on the object entered.
Additionally, in the example, we set \code{nsimu=100} for testing time considerations.
In reality, \code{nsimu} is typically set to 5000 to ensure the accuracy of the results.
}
\examples{
############design without late-onset outcomes################
nsimu <- 100; ncohort <- 12; cohortsize <- 3; init.level <- 1
p.true <- c(0.01, 0.02, 0.05, 0.20, 0.30, 0.50, 0.70)
target <- 0.2; prior.para = list(alp.prior = target, bet.prior = 1 - target)
## CFO design
CFOtrial <- CFO.simu(target, p.true, ncohort, init.level = 1, cohortsize = 3, design = 'CFO',
            prior.para)
plot(CFOtrial)
CFOoc <- CFO.oc (nsimu, design='CFO', target, p.true, ncohort, init.level, cohortsize,
                    tau = NaN, tte.para = NaN, accrual = NaN, accrual.dist = NaN, prior.para)
plot(CFOoc)
## aCFO design
aCFOtrial <- CFO.simu(target, p.true, ncohort, init.level = 1, cohortsize = 3, design = 'aCFO', 
              prior.para)
plot(aCFOtrial)
aCFOoc <- CFO.oc (nsimu, design = 'aCFO', target, p.true, ncohort, init.level, cohortsize,
                    tau = NaN, tte.para = NaN, accrual = NaN, accrual.dist = NaN, prior.para)
plot(aCFOoc)


##############design with late-onset outcomes################
tau <- 3; accrual <- 2; tte.para <- 0.5; accrual.dist <- 'unif'
## TITE-CFO design
 TITECFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                 accrual.dist, design = 'TITE-CFO', init.level = 1, 
                 prior.para = list(alp.prior = target, bet.prior = 1 - target))
plot(TITECFOtrial)
TITECFOoc <- CFO.oc (nsimu, design='TITE-CFO', target, p.true, ncohort, init.level, cohortsize,
                      tau, tte.para, accrual, accrual.dist, prior.para)
plot(TITECFOoc)
## TITE-aCFO design
TITEaCFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                accrual.dist, design = 'TITE-aCFO', init.level = 1, 
                prior.para = list(alp.prior = target, bet.prior = 1 - target))
plot(TITEaCFOtrial)
TITEaCFOoc <- CFO.oc (nsimu, design='TITE-aCFO', target, p.true, ncohort, init.level, cohortsize,
                      tau, tte.para, accrual, accrual.dist, prior.para)
plot(TITEaCFOoc)
## fCFO design
fCFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                accrual.dist, design = 'fCFO', init.level = 1, 
                prior.para = list(alp.prior = target, bet.prior = 1 - target))
plot(fCFOtrial)
fCFOoc <- CFO.oc (nsimu, design = 'fCFO', target, p.true, ncohort, init.level, cohortsize,
                      tau, tte.para, accrual, accrual.dist, prior.para)
plot(fCFOoc)
## f-aCFO design
faCFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, accrual, 
                accrual.dist, design = 'f-aCFO', init.level = 1, 
                prior.para = list(alp.prior = target, bet.prior = 1 - target))
plot(faCFOtrial)
faCFOoc <- CFO.oc (nsimu, design = 'f-aCFO', target, p.true, ncohort, init.level, cohortsize,
                      tau, tte.para, accrual, accrual.dist, prior.para)
plot(faCFOoc)

##############two-dim design with late-onset outcomes################
p.true <- matrix(c(0.05, 0.10, 0.15, 0.30, 0.45,
                   0.10, 0.15, 0.30, 0.45, 0.55,
                   0.15, 0.30, 0.45, 0.50, 0.60), 
                 nrow = 3, ncol = 5, byrow = TRUE)

## plot the single trail returned by CFO2d.simu()
CFO2dtrail <- CFO2d.simu(target = 0.3, p.true = p.true, ncohort = 20, cohortsize = 3)
plot(CFO2dtrail)

## plot the multiple trails returned by CFO2d.oc()
CFO2doc <- CFO2d.oc(target = 0.3, p.true = p.true, ncohort = 20, cohortsize = 3, 
            seeds = 1:100, nsimu = 100)
plot(CFO2doc)
}
\author{
Jialu Fang and Wenliang Wang
}
