% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summury.cfo.R
\name{summary.cfo}
\alias{summary.cfo}
\title{Generate descriptive summary for objects returned by other functions}
\usage{
\method{summary}{cfo}(object, ...)
}
\arguments{
\item{object}{the object returned by other functions.}

\item{...}{ignored arguments}
}
\value{
\code{summary()} prints the objects returned by other functions.
}
\description{
Generate descriptive summary for objects returned by other functions.
}
\details{
\code{summary()} prints the objects returned by other functions. Additionally, in the example,
we set \code{nsimu=100} for testing time considerations. In reality, \code{nsimu} is typically
set to 5000 to ensure the accuracy of the results.
}
\examples{
## settings for 1dCFO
nsimu <- 100; ncohort <- 12; cohortsize <- 3; init.level <- 1
p.true <- c(0.02, 0.05, 0.20, 0.28, 0.34, 0.40, 0.44)
target <- 0.2; prior.para=list(alp.prior=target, bet.prior=1-target)
tau <- 3; acrrual.rate <- 2; tte.para <- 0.5; accrual.dist <- 'unif'

## summarize the object returned by CFO.next()
decision <- CFO.next(target = 0.2, cys = c(0, 1, 0), cns = c(3, 6, 0), currdose = 3, prior.para)
summary(decision)

## summarize the object returned by CFO.simu()

aCFOtrial <- CFO.simu(target, p.true, ncohort, init.level = 1, cohortsize = 3, design = 'aCFO',
                      prior.para = list(alp.prior = target, bet.prior = 1 - target))
summary(aCFOtrial)

## summarize the object returned by lateonset.next()
enter.times<- c(0, 0.266, 0.638, 1.54, 2.48, 3.14, 3.32, 4.01, 4.39, 5.38, 5.76,
               6.54, 6.66, 6.93, 7.32, 7.65, 8.14, 8.74)
dlt.times<- c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0.995, 0, 0, 0, 0, 0, 0, 0, 2.58)
current.t<- 9.41
doses<-c(1, 1, 1, 2, 2, 2, 3, 3, 3, 4, 4, 4, 3, 3, 3, 4, 4, 4)
decision <- lateonset.next(target, p.true, currdose = 4, design = 'f-aCFO', tau = 3, enter.times,   
               dlt.times, current.t, doses, prior.para)
summary(decision)

## summarize the object returned by lateonset.simu()
faCFOtrial <- lateonset.simu (target, p.true, tau, cohortsize, ncohort, tte.para, acrrual.rate,
                accrual.dist, design = 'f-aCFO', init.level, prior.para)
summary(faCFOtrial)

## summarize the object returned by CFO.oc()
faCFOoc <- CFO.oc (nsimu, design = 'f-aCFO', target, p.true, ncohort, init.level, cohortsize,
                      tau, tte.para, acrrual.rate, accrual.dist, prior.para)
summary(faCFOoc)

## settings for 2dCFO
p.true <- matrix(c(0.05, 0.10, 0.15, 0.30, 0.45,
0.10, 0.15, 0.30, 0.45, 0.55,
0.15, 0.30, 0.45, 0.50, 0.60), 
nrow = 3, ncol = 5, byrow = TRUE)

cns <- matrix(c(3, 3, 0,
                0, 6, 0,
                0, 0, 0), 
              nrow = 3, ncol = 3, byrow = TRUE)
cys <- matrix(c(0, 1, 0,
                0, 2, 0,
                0, 0, 0), 
              nrow = 3, ncol = 3, byrow = TRUE)
currdose <- c(2,3)

## summarize the object returned by CFO2d.next()
decision <- CFO2d.next(0.3, cys, cns, currdose = currdose)
summary(decision)

## summarize the object returned by CFO2d.simu()
CFO2dtrail <- CFO2d.simu(target=0.3, p.true=p.true, ncohort = 20, cohortsize = 3)
summary(CFO2dtrail)

## summarize the object returned by CFO2d.oc()
CFO2doc <- CFO2d.oc(target=0.3, p.true=p.true, ncohort = 20, cohortsize = 3, nsimu = 100)
summary(CFO2doc)

}
\author{
Jialu Fang and Wenliang Wang
}
